EVENT GRID AND MODALS DONE
the bootstrap iso thing did not work turns out it wasn't the library I needed so I deleted all of bootstrap and went back and looked through the bulma documentation and figured out a way to make a COMPLETELY RESPONSIVE grid and I added all of the popups for it (javascript done- appearance not done). Now that this is done I am finally going to have a half hour nap before waking up for class. Fuck you.